+++ clang/docs/ReleaseNotes.rst
@@ -942,15 +942,15 @@ X86 Support
 
   * The options ``-m3dnow`` and ``-m3dnowa`` are no longer honored, and will emit a warning if used.
   * The macros ``__3dNOW__`` and ``__3dNOW_A__`` are no longer ever set by the compiler.
-  * The header ``<mm3dnow.h>`` still exists, but all of the the 3dNow
-    intrinsic functions have been removed: ``_m_femms``,
+  * The header ``<mm3dnow.h>`` is deprecated, and emits a warning if included.
+  * The 3dNow intrinsic functions have been removed: ``_m_femms``,
     ``_m_pavgusb``, ``_m_pf2id``, ``_m_pfacc``, ``_m_pfadd``,
     ``_m_pfcmpeq``, ``_m_pfcmpge``, ``_m_pfcmpgt``, ``_m_pfmax``,
     ``_m_pfmin``, ``_m_pfmul``, ``_m_pfrcp``, ``_m_pfrcpit1``,
     ``_m_pfrcpit2``, ``_m_pfrsqrt``, ``_m_pfrsqrtit1``, ``_m_pfsub``,
     ``_m_pfsubr``, ``_m_pi2fd``, ``_m_pmulhrw``, ``_m_pf2iw``,
     ``_m_pfnacc``, ``_m_pfpnacc``, ``_m_pi2fw``, ``_m_pswapdsf``,
-    ``_m_pswapdsi``
+    ``_m_pswapdsi``.
   * The compiler builtins corresponding to each of the above
     intrinsics have also been removed  (``__builtin_ia32_femms``, and so on).